+++ src/ui/views/GodMode.tsx
@@ -44,6 +44,7 @@ type Key =
 	| "challengeNoFreeAgents"
 	| "challengeNoRatings"
 	| "challengeNoTrades"
+	| "challengeLoseBestPlayer"
 	| "realPlayerDeterminism"
 	| "repeatSeason"
 	| "ties"
@@ -77,6 +78,8 @@ type Decoration = "currency" | "percent";
 type Values = { [key: string]: string | undefined };
 
 export const helpTexts = {
+	challengeLoseBestPlayer:
+		"At the end of the playoffs every season, the best player on your team will either retire (if he's a real player) or die a tragic death (if he's a random player).",
 	challengeNoDraftPicks:
 		"Your team will not be given any draft picks. You can still trade with other teams to acquire their picks.",
 	challengeNoFreeAgents:
@@ -506,6 +509,12 @@ export const options: {
 		category: "Challenge Modes",
 		key: "challengeNoRatings",
 		name: "No Visible Player Ratings",
+		type: "bool",
+	},
+	{
+		category: "Challenge Modes",
+		key: "challengeLoseBestPlayer",
+		name: "Lose Best Player",
 		type: "bool",
 	},
 	{